Crud sede bloque unidad

+++ src/containers/Admin/AdminSedes/Sedes.jsx
@@ -1,10 +1,12 @@
 import axios from "axios";
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
+import Swal from "sweetalert2";
 import { getSedes } from "../../../redux/sedes";
 import { Table, Button, message} from "antd";
 import useAuthorize from "../../../utils/authorization";
+import isValid from "../../../utils/specialChars";
+import './index.css'
 
 export default function Sedes() {
   const [form, setForm] = useState({});
@@ -20,6 +22,7 @@ export default function Sedes() {
 
   const handleSubmit = (e, id) => {
       e.preventDefault()
+      if (!isValid(form.nombre)) return message.error("No se permiten caracteres especiales")
     axios
       .post(`/api/sedes/${id}`, form)
       .then((res) => res.data)
@@ -30,10 +33,41 @@ export default function Sedes() {
     setForm({...form, [e.target.name]: e.target.value });
   };
 
+  const alertaEliminar = Swal.mixin({
+    buttonsStyling: true,
+  });
+
   const handleDelete = (id) => {
-    axios.delete(`/api/sedes/${id}`).then(() => dispatch(getSedes()));
-    message.success("Sede eliminada correctamente")
+    alertaEliminar
+      .fire({
+        title: "Estás seguro?",
+        text: "Si lo confirmas, no podrás deshacerlo!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Si, borrar!",
+        cancelButtonText: "No, cancelar!",
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          alertaEliminar.fire(
+            "Eliminado!",
+            "La sede fue eliminada correctamente.",
+            "success"
+          );
+          axios.delete(`/api/sedes/${id}`).then(() => dispatch(getSedes()));
+          message.success("Sede eliminada correctamente")
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          alertaEliminar.fire("Cancelado", "La sede está a salvo", "error");
+        }
+      });
   };
+
+
+
+
 
   useAuthorize(user, 1)
 
@@ -68,7 +102,7 @@ export default function Sedes() {
           >
             Modificar sede
           </Button>
-          <form onSubmit={(e) => handleSubmit(e, record.key)} id={`sedeForm${record.key}`} style={{display: 'none'}}>
+          <form className='sedesForm' onSubmit={(e) => handleSubmit(e, record.key)} id={`sedeForm${record.key}`} style={{display: 'none'}}>
                             <label htmlFor='nombre'>Nombre</label>
                             <input onChange={handleChange} type='text' name='nombre' placeholder={record.nombre}></input>
                             <label htmlFor='comunidadId'>Comunidad ID</label>
@@ -85,8 +119,7 @@ export default function Sedes() {
 
   return (
     <div>
-      {console.log(form)}
-      <Table dataSource={dataSource} columns={columns} pagination={false} />
+      <Table bordered dataSource={dataSource} columns={columns} pagination={false} />
     </div>
   );
 }
